report-management - rename actions to operations
