+++ src/lib/bark/operations/has-appointment.ts
